requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
google-generativeai==0.8.0
httpx==0.25.2
scipy==1.11.4
librosa==0.10.0
numpy==1.24.3
pydantic==2.7.0
pyngrok==7.0.0
sarvamai
